# openapi.yaml
openapi: "3.0.0"
info:
  title: VastProxy OpenAI-Compatible API
  version: "1.0.0"
servers:
  - url: /v1
paths:
  /chat/completions:
    post:
      operationId: createChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatCompletionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateChatCompletionResponse"
  /models:
    get:
      operationId: listModels
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListModelsResponse"
  /models/{model}:
    get:
      operationId: retrieveModel
      parameters:
        - in: path
          name: model
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
components:
  schemas:
    CreateChatCompletionRequest:
      type: object
      required:
        - model
        - messages
      properties:
        model:
          type: string
        messages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ChatCompletionRequestMessage"
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 1
          nullable: true
        top_p:
          type: number
          minimum: 0
          maximum: 1
          default: 1
          nullable: true
        n:
          type: integer
          minimum: 1
          maximum: 128
          default: 1
          nullable: true
        stream:
          type: boolean
          nullable: true
          default: false
        stop:
          nullable: true
          oneOf:
            - type: string
            - type: array
              items:
                type: string
              minItems: 1
              maxItems: 4
        max_tokens:
          type: integer
          nullable: true
        max_completion_tokens:
          type: integer
          nullable: true
        presence_penalty:
          type: number
          default: 0
          minimum: -2
          maximum: 2
          nullable: true
        frequency_penalty:
          type: number
          default: 0
          minimum: -2
          maximum: 2
          nullable: true
        logit_bias:
          type: object
          nullable: true
          additionalProperties:
            type: integer
        logprobs:
          type: boolean
          default: false
          nullable: true
        top_logprobs:
          type: integer
          minimum: 0
          maximum: 20
          nullable: true
        user:
          type: string
        seed:
          type: integer
          nullable: true
        tools:
          type: array
          items:
            $ref: "#/components/schemas/ChatCompletionTool"
        tool_choice:
          oneOf:
            - type: string
              enum:
                - none
                - auto
                - required
            - $ref: "#/components/schemas/ChatCompletionNamedToolChoice"
        parallel_tool_calls:
          type: boolean
          default: true
        stream_options:
          type: object
          nullable: true
          properties:
            include_usage:
              type: boolean
        response_format:
          type: object
          properties:
            type:
              type: string
              enum:
                - text
                - json_object
                - json_schema
            json_schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                schema:
                  type: object
                  additionalProperties: true
                strict:
                  type: boolean
                  nullable: true

    ChatCompletionRequestMessage:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum:
            - system
            - user
            - assistant
            - tool
        content:
          nullable: true
          oneOf:
            - type: string
            - type: array
              items:
                $ref: "#/components/schemas/ChatCompletionContentPart"
        name:
          type: string
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/ChatCompletionMessageToolCall"
        tool_call_id:
          type: string

    ChatCompletionContentPart:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - text
            - image_url
        text:
          type: string
        image_url:
          type: object
          properties:
            url:
              type: string
            detail:
              type: string
              enum:
                - auto
                - low
                - high

    ChatCompletionTool:
      type: object
      required:
        - type
        - function
      properties:
        type:
          type: string
          enum:
            - function
        function:
          $ref: "#/components/schemas/FunctionObject"

    FunctionObject:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        parameters:
          type: object
          additionalProperties: true
        strict:
          type: boolean
          nullable: true
          default: false

    ChatCompletionNamedToolChoice:
      type: object
      required:
        - type
        - function
      properties:
        type:
          type: string
          enum:
            - function
        function:
          type: object
          required:
            - name
          properties:
            name:
              type: string

    ChatCompletionMessageToolCall:
      type: object
      required:
        - id
        - type
        - function
      properties:
        id:
          type: string
        type:
          type: string
          enum:
            - function
        function:
          type: object
          required:
            - name
            - arguments
          properties:
            name:
              type: string
            arguments:
              type: string

    CreateChatCompletionResponse:
      type: object
      required:
        - id
        - object
        - created
        - model
        - choices
      properties:
        id:
          type: string
        object:
          type: string
          enum:
            - chat.completion
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChatCompletionChoice"
        usage:
          $ref: "#/components/schemas/CompletionUsage"
        system_fingerprint:
          type: string
        service_tier:
          type: string
          nullable: true

    ChatCompletionChoice:
      type: object
      required:
        - index
        - message
        - finish_reason
      properties:
        index:
          type: integer
        message:
          $ref: "#/components/schemas/ChatCompletionResponseMessage"
        finish_reason:
          type: string
          nullable: true
          enum:
            - stop
            - length
            - tool_calls
            - content_filter
        logprobs:
          nullable: true
          type: object
          properties:
            content:
              nullable: true
              type: array
              items:
                $ref: "#/components/schemas/ChatCompletionTokenLogprob"

    ChatCompletionResponseMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum:
            - assistant
        content:
          type: string
          nullable: true
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/ChatCompletionMessageToolCall"
        refusal:
          type: string
          nullable: true

    ChatCompletionTokenLogprob:
      type: object
      required:
        - token
        - logprob
        - bytes
        - top_logprobs
      properties:
        token:
          type: string
        logprob:
          type: number
        bytes:
          nullable: true
          type: array
          items:
            type: integer
        top_logprobs:
          type: array
          items:
            type: object
            required:
              - token
              - logprob
              - bytes
            properties:
              token:
                type: string
              logprob:
                type: number
              bytes:
                nullable: true
                type: array
                items:
                  type: integer

    CompletionUsage:
      type: object
      required:
        - prompt_tokens
        - completion_tokens
        - total_tokens
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
        completion_tokens_details:
          type: object
          properties:
            reasoning_tokens:
              type: integer
              default: 0
        prompt_tokens_details:
          type: object
          properties:
            cached_tokens:
              type: integer
              default: 0

    Model:
      type: object
      required:
        - id
        - object
        - created
        - owned_by
      properties:
        id:
          type: string
        object:
          type: string
          enum:
            - model
        created:
          type: integer
        owned_by:
          type: string

    ListModelsResponse:
      type: object
      required:
        - object
        - data
      properties:
        object:
          type: string
          enum:
            - list
        data:
          type: array
          items:
            $ref: "#/components/schemas/Model"